requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
pinecone
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
gunicorn==21.2.0
pyjwt==2.8.0
